natdis-detect-fe/natdis/requirements.txt
django
requests
gunicorn
pyjwt
